refactor(client): Replace relative shared imports with @zakapp/shared
- Update all React components, hooks, pages, and services
- Replace '../../../shared/src/types' with '@zakapp/shared'
- Replace '../../../../shared/src/types' with '@zakapp/shared'
- Standardize import paths across frontend codebase

Affected areas:
- Components (assets, dashboard, tracking, zakat)
- All custom hooks (analytics, payments, snapshots, etc.)
- All pages (Dashboard, Calculator, Profile, Settings)
- Services and utilities (API, PDF generator, chart formatters)
