+++ client/src/hooks/useSnapshots.ts
@@ -4,7 +4,7 @@
  */
 
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import type { YearlySnapshot } from '../../../shared/types/tracking';
+import type { YearlySnapshot } from '@zakapp/shared/types/tracking';
 
 const API_BASE_URL = process.env.REACT_APP_API_BASE_URL || 'http://localhost:3001/api';
