+++ client/src/hooks/useUpdateSnapshot.ts
@@ -4,7 +4,7 @@
  */
 
 import { useMutation, useQueryClient, UseMutationResult } from '@tanstack/react-query';
-import type { UpdateYearlySnapshotDto, YearlySnapshot } from '../../../shared/types/tracking';
+import type { UpdateYearlySnapshotDto, YearlySnapshot } from '@zakapp/shared/types/tracking';
 
 const API_BASE_URL = process.env.REACT_APP_API_BASE_URL || 'http://localhost:3001/api';
